requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
gTTS==2.5.4
idna==3.10
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
